Save statistics client side too.

Closes #345

See merge request grgrig/voshooter!220

+++ core/src/main/java/ee/taltech/voshooter/screens/EndScreen.java
@@ -23,7 +23,7 @@ public class EndScreen implements Screen {
     private double timeElapsed = 0;
 
     private int gameMode;
-    private int gameLength;
+    private double gameLength;
     private int playerCount;
     private int botAmount;
     private GameMap.MapType mapType;
